app/src/main/java/com/yo/android/chat/ui/fragments/UserChatFragment.java

+++ app/src/main/java/com/yo/android/chat/ui/fragments/UserChatFragment.java
@@ -162,6 +162,7 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
         mLog.e(TAG, "Firebase token reading from pref " + preferenceEndPoint.getStringPreference(Constants.FIREBASE_TOKEN));
         authReference = fireBaseHelper.authWithCustomToken(getActivity(), preferenceEndPoint.getStringPreference(Constants.FIREBASE_TOKEN));
         mToastFactory.showToast(getResources().getDisplayMetrics().density + "");
+        chatMessageArray = new ArrayList<>();
         setHasOptionsMenu(true);
     }
 
@@ -183,7 +184,7 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
         cameraView = (ImageView) view.findViewById(R.id.cameraView);
         chatText = (EditText) view.findViewById(R.id.chat_text);
         noChatAvailable = (TextView) view.findViewById(R.id.no_chat_text);
-        chatMessageArray = new ArrayList<>();
+
         chatMessageHashMap = new HashMap<>();
         userChatAdapter = new UserChatAdapter(getActivity(), preferenceEndPoint.getStringPreference(Constants.PHONE_NUMBER), roomType, mContactsSyncManager);
         listView.setAdapter(userChatAdapter);
@@ -251,7 +252,7 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
             listView.setTranscriptMode(ListView.TRANSCRIPT_MODE_NORMAL);
             listView.setStackFromBottom(false);
 
-            listView.setOnScrollChangeListener(new View.OnScrollChangeListener() {
+            /*listView.setOnScrollChangeListener(new View.OnScrollChangeListener() {
                 @Override
                 public void onScrollChange(View v, int scrollX, int scrollY, int oldScrollX, int oldScrollY) {
                     try {
@@ -263,7 +264,7 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
                         mLog.w("UserChat", e);
                     }
                 }
-            });
+            });*/
         } catch (NoClassDefFoundError e) {
             mLog.w("UserChat", e);
         }
@@ -546,6 +547,7 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
                 }
             });
 
+
         } catch (FirebaseException | NullPointerException e) {
             e.printStackTrace();
         }
@@ -618,11 +620,15 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
                 break;
             case ADD_IMAGE_CAPTURE:
                 try {
+                    if (data != null) {
+                        Uri targetUri = data.getData();
+                    }
                     String mPartyPicUri = mFileTemp.getPath();
                     String path = new CompressImage(getActivity()).compressImage(mPartyPicUri);
                     mFileTemp.delete();
                     updateChatWithLocalImage(path);
                 } catch (Exception e) {
+
                 }
                 break;
 
@@ -655,9 +661,14 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
                                 getActivity().runOnUiThread(new Runnable() {
                                     @Override
                                     public void run() {
-                                        String path = new CompressImage(getActivity()).compressImage(mFileTemp.getAbsolutePath());
-                                        mFileTemp.delete();
-                                        updateChatWithLocalImage(path);
+                                        if (mFileTemp != null && !isKb(mFileTemp.length())) {
+                                            String path = new CompressImage(getActivity()).compressImage(mFileTemp.getAbsolutePath());
+                                            mFileTemp.delete();
+                                            updateChatWithLocalImage(path);
+                                        } else {
+                                            updateChatWithLocalImage(mFileTemp.getAbsolutePath());
+                                        }
+
                                     }
                                 });
                             }
@@ -674,6 +685,11 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
             default:
                 break;
         }
+    }
+
+    private boolean isKb(long length) {
+        double size = length / 1024.0;
+        return size > 1 ? false : true;
     }
 
     public static void copyFile(String inputPath, String outputPath) {
@@ -801,6 +817,7 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
 
             ChatMessage chatMessage = dataSnapshot.getValue(ChatMessage.class);
             if (!chatMessageHashMap.keySet().contains(chatMessage.getMsgID())) {
+
                 chatMessageArray.add(chatMessage);
                 userChatAdapter.addItems(chatMessageArray);
                 listView.smoothScrollToPosition(userChatAdapter.getCount());
@@ -954,7 +971,6 @@ public class UserChatFragment extends BaseFragment implements View.OnClickListen
                     emojicon.getEmoji().length());
         }
     }
-
 
     private void changeEmojiKeyboardIcon(ImageView iconToBeChanged, int drawableResourceId) {
         iconToBeChanged.setImageResource(drawableResourceId);
